requirements.txt
PyYAML
pandas
scikit-learn
numpy
h5py
tqdm
tensorflow-cpu==2.9.0